fix(phase1): address fifth Codex review pass on PR #112
Seven more findings, one P1:

- LegacyAttendanceImportService (P1): a failed rollback no longer lets
  close() restore autoCommit. Flipping autoCommit false -> true on a
  connection whose transaction never rolled back implicitly commits
  it under JDBC, which could silently commit a supposedly-failed
  attendance import despite the endpoint answering 500.
- LegacyAttendanceImportService.analyze: broadened the catch from
  LegacyAttendanceImportException to the full RuntimeException
  category, matching PHP's catch (RuntimeException) and this method's
  own doc comment -- a DataAccessException from a dropped connection
  now maps to invalid_file_type instead of escaping to D-084's 500.
- LegacyAttendanceImportReader.sheetCode: (?U) on the whitespace
  collapse so Unicode whitespace (e.g. a non-breaking space) matches,
  same as PHP's /u modifier.
- LegacyScheduleCalendar.rowFromShift: the exception-note trim now
  uses LegacyValues.phpTrim, same form-feed class of bug as the other
  trim sites already fixed in this PR.
- LegacyWeeklyOffDays.phpIntCast: a numeric string that overflows a
  double (e.g. "1e9999") now casts to 0, matching PHP's documented
  NaN/Infinity-to-int behavior instead of Java's narrowing conversion,
  which clamps to Integer.MAX_VALUE.
- LegacyAttendanceImportReader.DATE/DATE_TIME: uuuu instead of yyyy,
  so a punch dated year 0000 formats as "0000" (proleptic year)
  instead of "0001" (year-of-era).
- LegacyAttendanceLocation.radiusMeters: kept at 64-bit width, so an
  unsigned radius_meters value above Integer.MAX_VALUE no longer
  wraps negative and silently falls back to the 200m default.

+++ backend/src/main/java/com/workin/legacy/attendance/location/LegacyAttendanceLocation.java
@@ -101,7 +101,7 @@ public class LegacyAttendanceLocation {
 
 		double distance = haversineMeters(latitude, longitude,
 				toDouble(branch.get("latitude")), toDouble(branch.get("longitude")));
-		int radius = radiusMeters(branch);
+		long radius = radiusMeters(branch);
 		if (distance > radius) {
 			throw outOfRange(distance, radius);
 		}
@@ -134,11 +134,11 @@ public class LegacyAttendanceLocation {
 		}
 
 		Double nearestDistance = null;
-		int nearestRadius = DEFAULT_RADIUS_METERS;
+		long nearestRadius = DEFAULT_RADIUS_METERS;
 		for (Map<String, Object> branch : branches) {
 			double distance = haversineMeters(latitude, longitude,
 					toDouble(branch.get("latitude")), toDouble(branch.get("longitude")));
-			int radius = radiusMeters(branch);
+			long radius = radiusMeters(branch);
 			if (distance <= radius) {
 				return;
 			}
@@ -270,10 +270,15 @@ public class LegacyAttendanceLocation {
 		return id > 0 ? id : null;
 	}
 
-	/** {@code branch_attendance_radius_meters()}: a non-positive radius becomes 200. */
-	public static int radiusMeters(Map<String, Object> branch) {
+	/**
+	 * {@code branch_attendance_radius_meters()}: a non-positive radius becomes
+	 * 200. Kept at 64-bit width: {@code radius_meters} is an unsigned database
+	 * column, so a value above {@link Integer#MAX_VALUE} is legitimate and
+	 * must not wrap negative and fall back to the default.
+	 */
+	public static long radiusMeters(Map<String, Object> branch) {
 		Object raw = branch == null ? null : branch.get("radius_meters");
-		int radius = raw == null ? DEFAULT_RADIUS_METERS : (int) LegacyValues.toPhpLong(raw);
+		long radius = raw == null ? DEFAULT_RADIUS_METERS : LegacyValues.toPhpLong(raw);
 		return radius > 0 ? radius : DEFAULT_RADIUS_METERS;
 	}
 
@@ -297,7 +302,7 @@ public class LegacyAttendanceLocation {
 	}
 
 	/** {@code fail_out_of_attendance_range()}: 400, with the distance rounded to whole metres. */
-	private static LegacyApiException outOfRange(double distanceMeters, int radiusMeters) {
+	private static LegacyApiException outOfRange(double distanceMeters, long radiusMeters) {
 		return new LegacyApiException(400, "out_of_range", null, Map.of(
 				"dist", String.valueOf((long) Math.round(distanceMeters)),
 				"radius", String.valueOf(radiusMeters)));
